Add files via upload
The Garment Industry is one of the key examples of the industrial globalization of this modern era. It is a
highly labour-intensive industry with lots of manual processes. it is highly desirable among the decision
makers in the garments industry to track, analyse and predict the productivity performance of the
working teams in their factories. Tried to predict the productivity.